Some more cleanup and removal
